move things in "the right place"
require base/js/namespace, or it might not be on the page when we access
IPython, hence random crash.

trigger configuration at load time, or IPython might not be initialized,
and it might crash.

Keep in mind thas JS is completly asynchrone, and it's not because
things works locally that they will always work.